+++ src/hexsample/tasks.py
@@ -166,13 +166,13 @@ class ReconstructionDefaults:
     num_neighbors: int = 2
     max_neighbors: int = -1
     pos_recon_algorithm: str = "centroid"
-    gain_map: Optional[np.ndarray] = None
-    eta_2pix_rad: float = 0.127
-    eta_2pix_pivot: float = 0.04
-    eta_3pix_rad0: float = 0.513
-    eta_3pix_rad1: float = 0.141
+    eta_2pix_rad_sigma: float = 0.127
+    eta_2pix_rad_pivot: float = 0.04
+    eta_3pix_rad_offset: float = 0.513
+    eta_3pix_rad_sigma: float = 0.141
     eta_3pix_rad_pivot: float = 0.05
-    eta_3pix_theta0: float = 0.104
+    eta_3pix_theta_sigma: float = 0.104
+    gain_map: Optional[np.ndarray] = None
 
 
 def reconstruct(
@@ -182,13 +182,13 @@ def reconstruct(
         num_neighbors: int = ReconstructionDefaults.num_neighbors,
         max_neighbors: int = ReconstructionDefaults.max_neighbors,
         pos_recon_algorithm: str = ReconstructionDefaults.pos_recon_algorithm,
-        gain_map: Optional[np.ndarray] = ReconstructionDefaults.gain_map,
-        eta_2pix_rad: float = ReconstructionDefaults.eta_2pix_rad,
-        eta_2pix_pivot: float = ReconstructionDefaults.eta_2pix_pivot,
-        eta_3pix_rad0: float = ReconstructionDefaults.eta_3pix_rad0,
-        eta_3pix_rad1: float = ReconstructionDefaults.eta_3pix_rad1,
+        eta_2pix_rad_sigma: float = ReconstructionDefaults.eta_2pix_rad_sigma,
+        eta_2pix_rad_pivot: float = ReconstructionDefaults.eta_2pix_rad_pivot,
+        eta_3pix_rad_offset: float = ReconstructionDefaults.eta_3pix_rad_offset,
+        eta_3pix_rad_sigma: float = ReconstructionDefaults.eta_3pix_rad_sigma,
         eta_3pix_rad_pivot: float = ReconstructionDefaults.eta_3pix_rad_pivot,
-        eta_3pix_theta0: float = ReconstructionDefaults.eta_3pix_theta0,
+        eta_3pix_theta_sigma: float = ReconstructionDefaults.eta_3pix_theta_sigma,
+        gain_map: Optional[np.ndarray] = ReconstructionDefaults.gain_map,
         header_kwargs: dict = None,
         ) -> str:
     """Run the reconstruction.
@@ -220,11 +220,26 @@ def reconstruct(
     pos_recon_algorithm : str
         The position reconstruction algorithm to use.
 
+    eta_2pix_rad_sigma : float
+        The sigma parameter for the radial component of the eta function for two pixel events.
+
+    eta_2pix_rad_pivot : float
+        The pivot parameter for the radial component of the eta function for two pixel events.
+
+    eta_3pix_rad_offset : float
+        The offset parameter for the radial component of the eta function for three pixel events.
+
+    eta_3pix_rad_sigma : float
+        The sigma parameter for the radial component of the eta function for three pixel events.
+
+    eta_3pix_rad_pivot : float
+        The pivot parameter for the radial component of the eta function for three pixel events.
+
+    eta_3pix_theta_sigma : float
+        The sigma parameter for the angular component of the eta function for three pixel events.
+
     gain_map : np.ndarray or None
         The gain map to use for the reconstruction. If None, no gain correction is applied.
-
-    eta_index : float
-        The eta index to use.
     """
     name, args = current_call()
     logger.info(f"Running {__name__}.{name} with arguments {args}...")
@@ -261,8 +276,21 @@ def reconstruct(
     # define it here because rectangular readout doesn't have a fixed number of neighbors, contrary
     # to the circular.
     effective_neighbors = max_neighbors if max_neighbors >= 0 else num_neighbors
+    # Create the dictionary with the reconstruction parameters to be passed to the clustering.
+    recon_pars = None
+    if pos_recon_algorithm == "eta":
+        recon_pars = dict(
+            eta_2pix_rad_sigma=eta_2pix_rad_sigma,
+            eta_2pix_rad_pivot=eta_2pix_rad_pivot,
+            eta_3pix_rad_offset=eta_3pix_rad_offset,
+            eta_3pix_rad_sigma=eta_3pix_rad_sigma,
+            eta_3pix_rad_pivot=eta_3pix_rad_pivot,
+            eta_3pix_theta_sigma=eta_3pix_theta_sigma,
+            pitch=header["pitch"]
+        )
     # Run the actual reconstruction.
-    clustering = ClusteringNN(readout, zero_sup_threshold, effective_neighbors)
+    clustering = ClusteringNN(readout, zero_sup_threshold, effective_neighbors,
+                              pos_recon_algorithm, recon_pars)
     output_file_path = input_file_path.replace(".h5", f"_{suffix}.h5")
     output_file = ReconOutputFile(output_file_path)
     if header_kwargs is not None:
@@ -278,9 +306,7 @@ def reconstruct(
         if cluster.size() in size:
             # Need to pass the recon method and other stuff as argument to ReconEvent
             args = event.trigger_id, event.timestamp(), event.livetime, cluster
-            recon_event = ReconEvent(*args, pos_recon_algorithm, readout.pitch,
-                                    eta_2pix_rad, eta_2pix_pivot, eta_3pix_rad0, eta_3pix_rad1,
-                                    eta_3pix_rad_pivot, eta_3pix_theta0)
+            recon_event = ReconEvent(*args)
             try:
                 mc_event = input_file.mc_event(i)
             except IndexError:
@@ -365,7 +391,8 @@ def calibrate(input_file_path: str,
     gain = CalibrationMatrixGain(header["num_cols"], header["num_rows"], energy, default_gain)
     noise = CalibrationMatrixNoise(header["num_cols"], header["num_rows"], default_noise)
     # Run the calibration.
-    clustering = ClusteringNN(readout, zero_sup_threshold=zero_sup_threshold, num_neighbors=6)
+    clustering = ClusteringNN(readout, zero_sup_threshold=zero_sup_threshold, num_neighbors=6,
+                              pos_recon_algorithm="centroid", recon_pars=None)
     for _, event in tqdm(enumerate(input_file)):
         try:
             cluster = clustering.run(event)
@@ -464,8 +491,17 @@ def display(
         raise RuntimeError(f"Unsupported readout mode: {readout_mode}")
     logger.info(f"Readout chip: {readout}")
     recon_defaults = ReconstructionDefaults
+    recon_pars = dict(
+        eta_2pix_rad_sigma=recon_defaults.eta_2pix_rad_sigma,
+        eta_2pix_rad_pivot=recon_defaults.eta_2pix_rad_pivot,
+        eta_3pix_rad_offset=recon_defaults.eta_3pix_rad_offset,
+        eta_3pix_rad_sigma=recon_defaults.eta_3pix_rad_sigma,
+        eta_3pix_rad_pivot=recon_defaults.eta_3pix_rad_pivot,
+        eta_3pix_theta_sigma=recon_defaults.eta_3pix_theta_sigma,
+        pitch=header["pitch"]
+    )
     _ = EventDisplay(input_file, readout, zero_sup_threshold=zero_sup_threshold,
-                     recon_defaults=recon_defaults)
+                     recon_pars=recon_pars)
     input_file.close()