+++ docs/release_notes.rst
@@ -3,6 +3,8 @@
 Release notes
 =============
 
+* Parameters for the eta function reconstruction renamed and packed into a
+  dictionary when passed around.
 * Implemented `SlantedEdgeResolution` class to calculate the estimate the resolution of the detector
   using the slanted edge method. This class calculates the edge spread function (ESF), the line
   spread function (LSF) and the modulation transfer function (MTF) for a slit beam.
@@ -13,6 +15,9 @@ Release notes
 
   - https://github.com/lucabaldini/hexsample/pull/102
   - https://github.com/lucabaldini/hexsample/pull/104
+  - https://github.com/lucabaldini/hexsample/pull/105
+  - https://github.com/lucabaldini/hexsample/issues/90
+  - https://github.com/lucabaldini/hexsample/issues/89
 
 
 Version 0.14.0 (2026-03-03)